Refactor the parser implementation
The current parser has bugs when multiple backslashes are used in
succession in quoted input. It also can be simplified quite a bit.

This refactoring does the following things:

1) Assigns the current character to a variable on every loop
iteration, so it doesn't need to constantly reference into the
buffer.

2) Makes the openQuote variable 3 valued, with the third value -1,
which indicates that at one point the word was quoted.  This is
reused to determine if the word NULL means nil (unquoted) or
"NULL" (quoted).

3) Adds a new flag escapeNext.  In quoted input, it means the
next character should be treated as a regular character instead
of a metacharacter.  Outside of quoted input, it means the last
entry was a subarray, and since it was already added to the array,
we shouldn't add the word.

4) Uses strncmp and rb_str_new instead of strcmp and rb_str_new2,
both of which take the word_index as the length, so we don't need
to worry about zero terminating the word.

5) Moves the empty array check outside the loop, so it doesn't
need to be checked in the loop.

6) Cleans up the if/else branching to be more consistent, instead
of mixing quoted and unquoted branches and trying to share a
common else.